Extract "core" plugin
