+++ lib/index.js
@@ -1,23 +1,7 @@
 // @flow
 import Options, { type OptionsFormat } from './options';
-import { schema, validateNode } from './validation';
 import { onEnter, onTab, onBackspace } from './handlers';
-import {
-    wrapInList,
-    unwrapList,
-    splitListItem,
-    increaseItemDepth,
-    decreaseItemDepth
-} from './changes';
-import {
-    getItemDepth,
-    isList,
-    isSelectionInList,
-    getCurrentItem,
-    getCurrentList,
-    getItemsAtRange,
-    getPreviousItem
-} from './utils';
+import core from './core';
 
 const KEY_ENTER = 'Enter';
 const KEY_TAB = 'Tab';
@@ -31,30 +15,12 @@ function EditList(
     opts: OptionsFormat = {}
 ): Object {
     opts = new Options(opts);
+    const corePlugin = core(opts);
 
     return {
-        onKeyDown: onKeyDown.bind(null, opts),
+        ...corePlugin,
 
-        schema: schema(opts),
-        validateNode: validateNode(opts),
-
-        utils: {
-            getCurrentItem: getCurrentItem.bind(null, opts),
-            getCurrentList: getCurrentList.bind(null, opts),
-            getItemDepth: getItemDepth.bind(null, opts),
-            getItemsAtRange: getItemsAtRange.bind(null, opts),
-            getPreviousItem: getPreviousItem.bind(null, opts),
-            isList: isList.bind(null, opts),
-            isSelectionInList: isSelectionInList.bind(null, opts)
-        },
-
-        changes: {
-            decreaseItemDepth: bindAndScopeChange(opts, decreaseItemDepth),
-            increaseItemDepth: bindAndScopeChange(opts, increaseItemDepth),
-            splitListItem: bindAndScopeChange(opts, splitListItem),
-            unwrapList: bindAndScopeChange(opts, unwrapList),
-            wrapInList: wrapInList.bind(null, opts)
-        }
+        onKeyDown: onKeyDown.bind(null, opts)
     };
 }
 
@@ -74,22 +40,6 @@ function onKeyDown(opts: Options, event, change, editor: *): void | any {
         default:
             return undefined;
     }
-}
-
-/**
- * Bind a change to given options, and scope it to act only inside a list
- */
-function bindAndScopeChange(opts: Options, fn: *): * {
-    return (change, ...args) => {
-        const { value } = change;
-
-        if (!isSelectionInList(opts, value)) {
-            return change;
-        }
-
-        // $FlowFixMe
-        return fn(...[opts, change].concat(args));
-    };
 }
 
 export default EditList;
